move color logic to new module
