+++ jp_gene_viz/dGraph.py
@@ -2,6 +2,8 @@ import numpy
 import heapq
 import json
 
+from color_scale import (clr, clr_check, weighted_color, color)
+
 def trim_leaves(Gin):
     Gout = WGraph()
     ew = Gin.edge_weights
@@ -78,18 +80,6 @@ def skeleton(Gin):
                     added.add(b)
     #Gout.node_weights = nw.copy()
     return Gout
-
-
-def clr(r, g, b):
-    result = numpy.array([r*1.0, g*1.0, b*1.0])
-    clr_check(result)
-    return result
-
-
-    
-def clr_check(clr):
-    assert max(clr) < 256
-    assert min(clr) >= 0
 
 
 class WGraph(object):
@@ -259,24 +249,10 @@ def draw_heat_map(canvas, a, dx, dy):
             canvas.rect(None, i*dx, j*dy, dx, dy, color(iclr))
     canvas.send_commands()
             
+
 def pos(x, y):
     return numpy.array([x*1.0, y*1.0])
 
-def weighted_color(maxclr, minclr, maxvalue, value):
-    assert value <= maxvalue
-    assert value >= 0
-    if maxvalue==0:
-        clr = minclr
-    else:
-        lm = value/float(maxvalue)
-        clr = (lm * maxclr) + ((1 - lm) * minclr)
-    return color(clr)
-
-def color(clr):
-    clr_check(clr)
-    ints = map(int, clr)
-    hexs = ["%02x" % x for x in ints]
-    return "#" + "".join(hexs)
 
 def towards(a, b, nonzero=True):
     diff = b - a
@@ -289,9 +265,11 @@ def towards(a, b, nonzero=True):
             return numpy.array([0,0])
     return diff/norm
 
+
 def orthogonal(v):
     [x,y] = v
     return pos(-y, x)
+
 
 def distance(a,b):
     return numpy.linalg.norm(b-a)
